+++ apps/be/src/modules/users/users.module.ts
@@ -3,12 +3,11 @@ import { UsersController } from "./users.controller";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { User } from "./entities/user.entity";
 import { UsersService } from "./users.service";
-import { JwtService } from "@nestjs/jwt";
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
   controllers: [UsersController],
-  providers: [UsersService, JwtService],
+  providers: [UsersService],
   exports: [UsersService],
 })
 export class UsersModule {}
